refactor: update plugin file structure
