+++ v-next/hardhat-viem/src/index.ts
@@ -5,7 +5,7 @@ import "./type-extensions.js";
 const hardhatPlugin: HardhatPlugin = {
   id: "hardhat-viem",
   hookHandlers: {
-    network: import.meta.resolve("./hook-handlers/network.js"),
+    network: import.meta.resolve("./internal/hook-handlers/network.js"),
   },
 };
